Bump some deps versions to be able to build with openssl v0.10
Both build in docker and on host fail with a fairly recent libssl
like 3.0.

This commit selectively bumps up some deps like native-tls and reqwest
to get the deps tree result only in openssl 0.10.x. This of course
lead to some package API changes and required a bit of modifications
to the source code in this project.

Unfortunately, rocket still requires nightly.

Built with rustc 1.46.0-nightly.